+++ docs/csv_subject_splitter/CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this gear are documented in this file.
 
 ## 1.0.2
 
-* Updates `uploader.JSONUploader` to query subject first and then add if it doesn't exist, instead of the other way around, since after any initial pipeline it is more likely for the subject to exist than not
+* Updates `ProjectAdaptor.add_subject` to return the subject if it already exists
+* Updates `uploader.JSONUploader` to remove the `allow_updates` parameter - will always update instead
 
 ## 1.0.1